pyproject.toml
[project]
name = "testa-de-ferro"
version = "0.1.0"
description = "Plataforma para identificar empresas com perfil suspeito que vendem ao governo federal"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "duckdb>=1.1,<2",
    "polars>=1.18,<2",
    "pydantic>=2.10,<3",
    "httpx>=0.28,<1",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-cov>=6.0,<7",
    "pytest-asyncio>=0.24,<1",
    "mypy>=1.13,<2",
    "ruff>=0.8,<1",
    "pre-commit>=4.0,<5",
    "pip-audit>=2.7,<3",
    "detect-secrets>=1.5,<2",
]
pdf = [
    "weasyprint>=63,<64",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "S",    # bandit (security)
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "T20",  # flake8-print (no print in prod code)
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # assert é OK em testes

[tool.setuptools.packages.find]
include = ["api*", "pipeline*"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
